This commit updates the DAGMaker and SEdiff setups to run in SL7.
Co-authored-by: Robert Morgan <robert.morgan@wisc.edu>

The scripts BLISS-expCalib_Y3apass-old.py and run_SEproc.py were updated to use SL7 packages only. However, SEdiff copies versions of them in from a tarball on the DES machine; the versions in the repo do not get used. These updated scripts must replace the scripts in the tarball.